requirements.txt
python-telegram-bot==21.11
APScheduler==3.10.4
PyYAML==6.0.2
requests==2.32.3
python-dateutil==2.9.0.post0
twikit
setuptools<70.0.0
